# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Fsirst API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/competitions": {
            "post": {
                "tags": [
                    "대회"
                ],
                "description": "대회 생성(팀의 리더만 대회를 생성 할 수 있음)",
                "operationId": "1bd0e57d4ea073be465faec070f45819",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tid": {
                                        "description": "호스트 팀아이디",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "kind": {
                                        "description": "대회 종목(1:풋살, 2:축구)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "type": {
                                        "description": "대회 타입(1:리그, 2:컵)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "region": {
                                        "description": "지역",
                                        "type": "string",
                                        "example": "서울시 노원구"
                                    },
                                    "limit_team ": {
                                        "description": "팀제한수",
                                        "type": "string",
                                        "example": "6"
                                    },
                                    "person_vs ": {
                                        "description": "경기인원",
                                        "type": "string",
                                        "example": "5:5,11:11 형태로 구분"
                                    },
                                    "regist_sdate": {
                                        "description": "모집시작일",
                                        "type": "string",
                                        "example": "2024-01-01 00:00:00"
                                    },
                                    "event_sdate": {
                                        "description": "경기(이벤트)시작일",
                                        "type": "string",
                                        "example": "2024-01-30 00:00:00"
                                    },
                                    "event_edate": {
                                        "description": "경기(이벤트)마감일",
                                        "type": "string",
                                        "example": "2024-02-28 00:00:00"
                                    },
                                    "frequency": {
                                        "description": "경기빈도",
                                        "type": "string",
                                        "example": "주1회, 주2회, 월5회,..."
                                    },
                                    "yoil": {
                                        "description": "경기선호요일",
                                        "type": "string",
                                        "example": "0,1,2 (0:일,1:월,2:화 콤마 형태로 구분)"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions?type=1&sorting": {
            "get": {
                "tags": [
                    "대회"
                ],
                "description": "대회 리스트 불러오기 [type=>0:리그,1:컵] [sorting =>pre:진행전,ing:진행중,end:종료된] [per_page=>페이지 당 레코드 갯수] [page =>보여줄 페이지]",
                "operationId": "d056644719313e2bceed8a24f7efebdd",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/search": {
            "post": {
                "tags": [
                    "대회"
                ],
                "description": "대회 리스트 불러오기 검색 ",
                "operationId": "78c2be161c404439dab2300f20c20f62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "description": "대회 검색(경기 title)",
                                        "type": "string",
                                        "example": "짧은"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/detail/{$cid}": {
            "get": {
                "tags": [
                    "대회"
                ],
                "description": "대회 상세 불러오기",
                "operationId": "f74733241bf31da67c0e93cc7316d961",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/{$cid}": {
            "patch": {
                "tags": [
                    "대회"
                ],
                "description": "대회 삭제 (display_yn='n') ",
                "operationId": "a651b9ad1e8e1d0d8a2f6732fddd6499",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/{cid}": {
            "post": {
                "tags": [
                    "대회"
                ],
                "description": "대회 수정",
                "operationId": "9549305136999741135cf7c08d396908",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kind": {
                                        "description": "대회 종목(1:풋살, 2:축구)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "type": {
                                        "description": "대회 타입(1:리그, 2:컵)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목"
                                    },
                                    "contents": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "region": {
                                        "description": "지역",
                                        "type": "string",
                                        "example": "서울시 노원구"
                                    },
                                    "limit_team ": {
                                        "description": "팀제한수",
                                        "type": "string",
                                        "example": "6"
                                    },
                                    "person_vs ": {
                                        "description": "경기인원",
                                        "type": "string",
                                        "example": "5:5,11:11 형태로 구분"
                                    },
                                    "regist_sdate": {
                                        "description": "모집시작일",
                                        "type": "string",
                                        "example": "2024-01-01 00:00:00"
                                    },
                                    "event_sdate": {
                                        "description": "경기(이벤트)시작일",
                                        "type": "string",
                                        "example": "2024-01-30 00:00:00"
                                    },
                                    "event_edate": {
                                        "description": "경기(이벤트)마감일",
                                        "type": "string",
                                        "example": "2024-02-28 00:00:00"
                                    },
                                    "frequency": {
                                        "description": "경기빈도",
                                        "type": "string",
                                        "example": "주1회, 주2회, 월5회,..."
                                    },
                                    "yoil": {
                                        "description": "경기선호요일",
                                        "type": "string",
                                        "example": "0,1,2 (0:일,1:월,2:화 콤마 형태로 구분)"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/apply/{cid}": {
            "post": {
                "tags": [
                    "대회"
                ],
                "description": "대회 참여 신청",
                "operationId": "c6521a2ae14ec2ca7eea1715abc0f41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cid": {
                                        "description": "참여 경기의 sid",
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/competitions/start/{cid}": {
            "post": {
                "tags": [
                    "대회"
                ],
                "description": "대회 경기 시작, 대회 state값 변경 및 매치 생성",
                "operationId": "cf7fea8e5fa7d61d9d765222af8bde42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cid": {
                                        "description": "참여 경기의 sid",
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
...
    }
}